test: address PR #94 review feedback
- Add per-package asModel helpers (configeditor, menueditor, stringeditor)
  so tea.Model type assertions fail the test instead of panicking; use
  them at every Update call site.
- configeditor: assert the clean-exit command actually yields tea.QuitMsg,
  not just any non-nil command.
- menueditor: split model_test.go (335 lines) below the 300-line guideline
  by extracting shared helpers to harness_test.go and the view smoke test
  to view_test.go.
- stringeditor: add doc comments to exported test functions; assert the
  "»" overflow marker is present in truncated RenderColorString output.
- menu: strengthen the CP437 displayFile subtest to compare the full raw
  output (including a high CP437 byte) against the pipe-code-expanded
  source, proving bytes pass through without UTF-8 translation.
